+++ packages/account-react/src/types/consent.ts
@@ -7,17 +7,3 @@ export type DecodedToken = {
   c?: ConsentClaims;
   [key: string]: unknown;
 };
-
-export type ConsentResponse = {
-  id: string;
-  content: {
-    result?: {
-      consentAds: boolean;
-      consentData: boolean;
-    };
-    error?: {
-      message: string;
-      code: number;
-    };
-  };
-};