+++ packages/account-api-sdk/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @sophon-labs/account-api-sdk
 
+## 3.0.0-next.34
+
+### Patch Changes
+
+- Refactoring
+- Updated dependencies
+  - @sophon-labs/account-core@3.0.0-next.34
+
+## 3.0.0-next.33
+
+### Patch Changes
+
+- .2
+- Updated dependencies
+  - @sophon-labs/account-core@3.0.0-next.33
+
+## 3.0.0-next.32
+
+### Patch Changes
+
+- .
+- Updated dependencies
+  - @sophon-labs/account-core@3.0.0-next.32
+
+## 3.0.0-next.31
+
+### Patch Changes
+
+- Updating viem extension
+- Updated dependencies
+  - @sophon-labs/account-core@3.0.0-next.31
+
+## 3.0.0-next.30
+
+### Patch Changes
+
+- Adding hooks and viem support for session request
+- Updated dependencies
+  - @sophon-labs/account-core@3.0.0-next.30
+
 ## 3.0.0-next.29
 
 ### Patch Changes